+++ ext/EMGExt/EMGExt.jl
@@ -15,5 +15,6 @@ const TS = TimeStruct
 
 include("compliance_access_fun.jl")
 include("compliance_ele.jl")
+include("mwe_case.jl")
 
 end