+++ NEWS.md
@@ -8,6 +8,6 @@ Initial version of the package:
   * Checks highlight whether a `Node` or `Link` can work with existing functions or have a new method.
   * Checks mostly provide warnings, as errors are dependent on the chosen functions from `EnergyModelsBase` or `EnergyModelsGeography`.
 
-* Inclusion of simple test case for `Node`
-  * Test case can be used to see whether the couplings in `EnergyModelsBase` are working.
+* Inclusion of simple test case for `Node` and `TransmissionMode`:
+  * Test case can be used to see whether the couplings in `EnergyModelsBase` `EnergyModelsGeography` are working.
   * Tests that capacity is utilized.